[arp]: Fix for single ToR testbeds
